+++ ui/litellm-dashboard/src/app/(dashboard)/models-and-endpoints/panels/LlmCredentialsPanel.tsx
@@ -4,7 +4,7 @@ import { Form } from "antd";
 import CredentialsPanel from "@/components/model_add/CredentialsPanel";
 import { vertexCredentialsUploadProps } from "@/app/(dashboard)/models-and-endpoints/vertexCredentialsUpload";
 
-export default function LlmCredentialsPage() {
+export default function LlmCredentialsPanel() {
   const [form] = Form.useForm();
   return <CredentialsPanel uploadProps={vertexCredentialsUploadProps(form)} />;
 }
