+++ ui/litellm-dashboard/src/app/(dashboard)/models-and-endpoints/panels/PassThroughPanel.tsx
@@ -3,7 +3,7 @@
 import PassThroughSettings from "@/components/PassThroughSettings/PassThroughSettings";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
-export default function PassThroughPage() {
+export default function PassThroughPanel() {
   const { accessToken, userRole, userId: userID, premiumUser } = useAuthorized();
   return (
     <PassThroughSettings accessToken={accessToken} userRole={userRole} userID={userID} premiumUser={premiumUser} />
